# public/swagger/swagger_1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger LinkeBin",
        "description": "The Linkebin api's documentation. You can find out more about Swagger at [http://swagger.io](http://swagger.io) and at [https://github.com/zircote/swagger-php](https://github.com/zircote/swagger-php)",
        "termsOfService": "https://smartbear.com/terms-of-use/",
        "contact": {
            "email": "contact@linkebin.com"
        },
        "license": {
            "name": "LinkeBin"
        },
        "version": "4.1.0"
    },
    "servers": [
        {
            "url": "linkebin.swagger.io",
            "description": "Api server"
        }
    ],
    "paths": {
        "/bin": {
            "get": {
                "tags": [
                    "bin"
                ],
                "summary": "Index of bin",
                "operationId": "App\\Controller\\BinController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Welcome to your new controller!"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/bins/getAllBins": {
            "get": {
                "tags": [
                    "bin"
                ],
                "summary": "Display all bins",
                "operationId": "displayBins",
                "parameters": [
                    {
                        "name": "Adress",
                        "in": "path",
                        "description": "Adress of bin",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "City",
                        "in": "path",
                        "description": "City of bin",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Coord",
                        "in": "path",
                        "description": "Coordonates of bin",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bin"
                                },
                                "example": "Bin on map"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/AddCityBins/{id}": {
            "post": {
                "tags": [
                    "bin"
                ],
                "summary": "Add all bins",
                "operationId": "createBins",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "query",
                        "description": "ID of bin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Url",
                        "in": "header",
                        "description": "Url of api",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bin"
                                },
                                "example": "\r\n     *         id: 0,\r\n     *         coords: 92.7667 87.4376,\r\n     *         city: Toulouse,\r\n     *         adress: 402 rue des tilleuls,\r\n     *         is_enable: true,\r\n     *         created_at: 2020-01-20T09:47:53.086Z,\r\n     *         updated_at: 2020-01-20T09:47:53.086Z,\r\n     *         user_bin: 42,\r\n     *         bin_historics: 50,\r\n     *         cityBins: 64"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Url doesn't exist",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "URL Doesn't Exist"
                            }
                        }
                    },
                    "default": {
                        "description": "Bin already here",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Bins déja présentes dans la bases de Donnés"
                            }
                        }
                    }
                }
            }
        },
        "/city": {
            "get": {
                "tags": [
                    "city"
                ],
                "summary": "Index of city",
                "operationId": "App\\Controller\\CityController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Welcome to your new controller!"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/addCity": {
            "post": {
                "tags": [
                    "city"
                ],
                "summary": "Create a city",
                "operationId": "createCities",
                "parameters": [
                    {
                        "name": "Link",
                        "in": "header",
                        "description": "Link of json file",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/city"
                                },
                                "example": "\r\n     *         id: 05,\r\n     *         file_json: [ C:\\Users\\17359\\PhpstormProjects\\LinkeBinAPI\\Bornes_a_verre(1).json ],\r\n     *         name: Toulouse,\r\n     *         region: Occitanie,\r\n     *         departement: Haute-Garonne,\r\n     *         cityBins: 109\r\n     *         "
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/report/historic": {
            "get": {
                "tags": [
                    "reporthistoric"
                ],
                "summary": "Index of report historic",
                "operationId": "App\\Controller\\ReportHistoricController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Welcome to your new controller!"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/AddReportHistoric/{idBin}/{idUser}": {
            "post": {
                "tags": [
                    "reporthistoric"
                ],
                "summary": "Add report historic",
                "operationId": "add Report Historic",
                "parameters": [
                    {
                        "name": "idBin",
                        "in": "path",
                        "description": "ID of bin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idUser",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/reporthistoric"
                                },
                                "example": "\r\n     *         id: 50,\r\n     *         uuid_users_bin: 05,\r\n     *         created_at: 2020-01-20T09:47:53.086Z,\r\n     *         degradation: true,\r\n     *         bin_full: true,\r\n     *         missing: false"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Index of users",
                "operationId": "App\\Controller\\UsersController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Welcome to your new controller!"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/AddUser/{id}": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Add all users",
                "operationId": "createUsers",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Token pour l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                },
                                "example": "\r\n     *         id: 29,\r\n     *         user_id: 05,\r\n     *         created_at: 2020-01-20T09:47:53.086Z,\r\n     *         updated_at: 2020-01-20T09:47:53.086Z,\r\n     *         token: dsqsdqsd,\r\n     *         is_enable: true,\r\n     *         user_bin: 64"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "bin": {
                "required": [
                    "id",
                    "coords",
                    "city",
                    "adress",
                    "is_enable",
                    "created_at",
                    "updated_at",
                    "user_bin",
                    "bin_historics",
                    "cityBins"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "coords": {
                        "type": "integer"
                    },
                    "city": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "adress": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "is_enable": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user_bin": {
                        "type": "integer"
                    },
                    "bin_historics": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "cityBins": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "type": "object"
            },
            "bin_historic": {
                "required": [
                    "id",
                    "uuid_bin",
                    "created_at",
                    "pickup_at",
                    "empty"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uuid_bin": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "pickup_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "empty": {
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "city": {
                "required": [
                    "id",
                    "file_json",
                    "name",
                    "region",
                    "departement",
                    "cityBins"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "file_json": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "region": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "departement": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "cityBins": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "citybin": {
                "required": [
                    "id",
                    "uuid_bin",
                    "uuid_city"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uuid_bin": {
                        "type": "integer"
                    },
                    "uuid_city": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "reporthistoric": {
                "required": [
                    "id",
                    "uuid_users_bin",
                    "created_at",
                    "degradation",
                    "bin_full",
                    "missing"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uuid_users_bin": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "degradation": {
                        "type": "boolean"
                    },
                    "bin_full": {
                        "type": "boolean"
                    },
                    "missing": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "users": {
                "required": [
                    "id",
                    "role",
                    "name",
                    "created_at",
                    "updated_at",
                    "adress",
                    "password",
                    "token",
                    "mail",
                    "is_enable",
                    "user_bin"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "token": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "is_enable": {
                        "type": "boolean"
                    },
                    "user_bin": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "usersbin": {
                "required": [
                    "id",
                    "uuid_bin",
                    "uuid_user",
                    "created_at",
                    "report_full",
                    "reportHistoric"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uuid_bin": {
                        "type": "integer",
                        "nullable": false
                    },
                    "uuid_user": {
                        "type": "integer",
                        "nullable": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "report_full": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "reportHistoric": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}